+++ under-utils-http/src/main/java/com/undernine/utils/http/client/package-info.java
@@ -1,10 +1,10 @@
 /**
- * HTTP 客户端实现包
+ * HTTP 请求执行器包
  * <p>
- * 包含不同 HTTP 客户端的具体实现。
+ * 包含不同底层 HTTP 客户端的执行器实现，对外承载 Under-Utils 自身的请求模型。
  * </p>
  * <ul>
- *   <li>{@link com.undernine.utils.http.client.OkHttpClient} - 基于 OkHttp 的实现（默认）</li>
+ *   <li>{@link com.undernine.utils.http.client.OkHttpRequestExecutor} - 基于 OkHttp 的请求执行器（默认）</li>
  * </ul>
  *
  * @author deng